chore: separate attention meta into another header file
